#342 Remade attachment_anonymization management command

+++ chcemvediet/apps/anonymization/management/commands/attachment_anonymization.py
@@ -5,51 +5,59 @@ from django.core.files.base import ContentFile
 from django.core.management.base import BaseCommand, CommandError
 
 from poleno.attachments.models import Attachment
+from poleno.utils.misc import squeeze
 from chcemvediet.apps.anonymization.models import AttachmentFinalization
 from chcemvediet.apps.inforequests.models import Action
 
 
 class Command(BaseCommand):
-    help = u'Creates AttachmentFinalization for the specified Attachment.'
+    args = u'attachment [file]'
+    help = squeeze(u"""
+            Creates anonymization for the specified Attachment. By default, the file path is read
+            from stdin. File path can be explicitly passed as an command argument. Anonymization
+            created this way will be marked as successful. Only one successful anonymization can be
+            assigned to the Attachment.
+            """)
 
     option_list = BaseCommand.option_list + (
-        make_option(u'-f', u'--file',
-                    dest=u'filename',
-                    help=u'define file path'),
         make_option(u'--content_type',
-                    help=u'define content type of file',
+                    help=squeeze(u"""
+                            Content type of file, e.g. "application/pdf". Automatically computed if
+                            not specified.
+                            """)
                     ),
         make_option(u'--debug',
                     default=u'',
-                    help=u'add debug message'
+                    help=u'Debug message to the newly created anonymization. Empty by default.'
                     ),
         make_option(u'--force',
                     action=u'store_true',
-                    help=u'overwrite an existing successful AttachmentFinalization'
+                    help=squeeze(u"""
+                            The command refuses to anonymize attachment if a successful
+                            anonymization already exists. This flag disables this check. Deletes all
+                            existing successful anonymizations and creates new one. Unsuccessful
+                            anonymizations will stay unaffected.
+                            """)
                     ),
     )
 
-    def usage(self, subcommand=None):
-        return u'manage.py attachment_anonymization attachment [options] [file]'
-
     def handle(self, *args, **options):
-        try:
-            filename = options[u'filename'] or args[1]
-        except IndexError:
-            self.stdout.write(u'Usage: {}'.format(self.usage()))
-            self.stdout.write(
-                u'Try "manage.py attachment_anonymization --help" for more information.')
-            return 1
-
+        if not args:
+            raise CommandError(u'attachment_anonymization takes at least 1 argument (0 given).')
+        if len(args) > 2:
+            raise CommandError(
+                u'attachment_anonymization takes at most 2 arguments ({} given).'.format(len(args))
+                )
+        pk = args[0]
+        filename = args[1] if len(args) == 2 else raw_input(u'File:')
+        if not filename:
+            raise CommandError(u'Missing source file name.')
         with open(filename, u'rb') as file:
             try:
-                attachment = (Attachment.objects
-                              .attached_to(Action)
-                              .not_normalized()
-                              .get(pk=int(args[0])))
+                attachment = Attachment.objects.attached_to(Action).get(pk=pk)
                 attachments_finalization = (AttachmentFinalization.objects
-                                 .filter(attachment=attachment)
-                                 .successful())
+                                            .filter(attachment=attachment)
+                                            .successful())
                 if options[u'force'] or not attachments_finalization:
                     attachments_finalization.delete()
                     content = file.read()
@@ -61,5 +69,10 @@ class Command(BaseCommand):
                         content_type=options[u'content_type'] or content_type,
                         debug=options[u'debug'],
                     )
+                else:
+                    raise CommandError(squeeze(u"""
+                            Anonymization files already exist. Use the --force option to overwrite
+                            them.
+                            """))
             except Attachment.DoesNotExist:
-                raise CommandError(u'Attachment {} does not exist'.format(args[0]))
+                raise CommandError(u'Attachment instance with pk {} does not exist.'.format(pk))
